+++ Modules/Intent.Modules.AspNetCore.Controllers/release-notes.md
@@ -1,3 +1,8 @@
+### Version 4.0.2
+
+- Fixed: Errors thrown when applying `application/json` to the Return Type Mediatype the Http Settings of an operation.
+- Adds `[Produces(MediaTypeNames.Application.Json)]` when explicitly specified in the Http Settings.
+
 ### Version 4.0.1
 
 - Fixed: Applying unsecured to a service would have no effect, it would need to be applied to individual operations.